# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "TriCore Agent API"
  description: "三核融合智能体 REST API - 意识感(白龙马) + 执行力(龙虾) + 成长性(爱马仕)"
  version: "2.4.0"
  contact:
    name: "TriCore Team"
  license:
    name: "MIT"
servers:
  - url: "http://localhost:3721"
    description: "本地开发服务器"
  - url: "https://api.tricore.example.com"
    description: "生产服务器"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    MessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: "消息内容"
        userId:
          type: string
          description: "用户ID"
        channel:
          type: string
          enum: [api, social, voice]
          default: api
        priority:
          type: integer
          minimum: 0
          maximum: 100
    MessageResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [queued, processing, completed, failed]
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/message/send:
    post:
      tags: [Message]
      summary: "发送消息"
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageRequest"
      responses:
        "200":
          description: "消息已接收"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...